Replace rule-based scoring with LightGBM ML pipeline (0-10 scale)
- Wire in trained LightGBM model (backend/models/regime_lgbm.txt) through
  a new model_service.py with batch + single-row prediction and sigmoid
  fallback; ensure_model_trained() called on app startup
- Replace scoring_engine.py: sub-scores remain integer flag counts (0-3/4),
  but total_score is now a continuous float 0-10 from the model
- Update regime_service.py band thresholds to 0-10 float scale
  (Risk-On <2.5, Neutral <5, Risk-Off <7.5, Crisis <=10)
- Fix FedWatch hook in regime_pipeline.py to nudge float total_score
  directly (±0.7) instead of recomputing raw integer sum
- Update schema/regime.py and regime_routes.py: total_score typed as float
- Refresh FRED macro data to include international yield series
  (boj_10y_yield, ecb_policy_rate, uk_10y_gilt_yield) that were missing
- Dashboard: remove static hardcoded icon from ScoreCards hero cards;
  remove non-functional Risk Gauge widget; collapse grid to 2 columns
- GlobalMacroTab: add empty-state fallback for charts with no data
- PlaybookTab: show score as X.X/10; fix raw -> arrow JSX parse error
- All score displays updated to reflect 0-10 ML scale

+++ backend/schema/regime.py
@@ -134,7 +134,7 @@ class RegimeCurrentResponse(BaseModel):
     regime_color: str
     risk_level: int
     probability: float
-    total_score: int
+    total_score: float
     breakdown: BreakdownResponse
     allocation: AllocationWeights
     etf_mapping: EtfMapping
@@ -150,7 +150,7 @@ class RegimeHistoryItem(BaseModel):
     regime_color: str
     risk_level: int
     probability: float
-    total_score: int
+    total_score: float
     growth_score: int
     inflation_score: int
     financial_score: int
@@ -204,5 +204,5 @@ class TransitionsResponse(BaseModel):
 # ─── scores endpoint ─────────────────────────────────────────────────
 
 class ScoresResponse(BaseModel):
-    total_score: int
+    total_score: float
     breakdown: BreakdownResponse
